requirements.txt
apache-airflow
flask
apache-airflow-providers-snowflake